No more warnings on x64.  - Used ptrdiff_t instead of int for index storage in luakey;  - Used ptrdiff_t instead of size_t everywhere.  - The call to lua_pcall casts the arguments count from size_t to int, because    Lua uses an int internally. It makes sense because lua_Integer is defined to ptrdiff_t.

+++ src/luadataimpl.cc
@@ -119,7 +119,7 @@ inline void luadataimpl::callfunction(const std::vector<luaarg> &args) {
 		}
 	}
 
-	lua_pcall(L, args.size(), 1, 0);
+	lua_pcall(L, static_cast<int>(args.size()), 1, 0);
 }
 
 double luadataimpl::getdoublefromstack(const double &defaultValue) {
@@ -291,12 +291,12 @@ luatype luadataimpl::type(const luapath &valuepath) {
 	return type;
 }
 
-std::size_t luadataimpl::tablelen(const luapath &valuepath) {
+std::ptrdiff_t luadataimpl::tablelen(const luapath &valuepath) {
 	// Gets the value at the front of the stack.
 	getpath(valuepath);
 
 	// Gets the table length.
-	std::size_t len = 0;
+	std::ptrdiff_t len = 0;
 	if(lua_istable(L, -1)) {
 		len = luaL_len(L, -1);
 	}
